chore: move snapshot to models and make pub
