+++ flipt-evaluation/src/models/mod.rs
@@ -1,2 +1,3 @@
 pub mod flipt;
+pub mod snapshot;
 pub mod source;
